requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Neo4j
neo4j==5.16.0

# Redis
redis==5.0.1

# Background Jobs
celery==5.3.4
flower==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI/ML
openai==1.10.0
langchain==0.1.6
langchain-openai==0.0.5
langgraph==0.0.20

# Web Scraping
playwright==1.41.0
beautifulsoup4==4.12.3
httpx==0.26.0
fake-useragent==1.4.0
aiolimiter==1.1.0
tenacity==8.2.3
dateparser==1.2.0

# Caching & Rate Limiting
slowapi==0.1.9

# Monitoring
sentry-sdk[fastapi]==1.40.0
python-json-logger==2.0.7

# Performance
psutil==5.9.8

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
locust==2.20.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
